ref(replay): Refactor eventBuffer into separate files
