+++ packages/replay/src/eventBuffer/index.ts
@@ -1,0 +1,28 @@
+import { logger } from '@sentry/utils';
+
+import type { EventBuffer } from '../types';
+import workerString from '../worker/worker.js';
+import { EventBufferArray } from './EventBufferArray';
+import { EventBufferProxy } from './EventBufferProxy';
+
+interface CreateEventBufferParams {
+  useCompression: boolean;
+}
+
+/**
+ * Create an event buffer for replays.
+ */
+export function createEventBuffer({ useCompression }: CreateEventBufferParams): EventBuffer {
+  // eslint-disable-next-line no-restricted-globals
+  if (useCompression && window.Worker) {
+    const workerBlob = new Blob([workerString]);
+    const workerUrl = URL.createObjectURL(workerBlob);
+
+    __DEBUG_BUILD__ && logger.log('[Replay] Using compression worker');
+    const worker = new Worker(workerUrl);
+    return new EventBufferProxy(worker);
+  }
+
+  __DEBUG_BUILD__ && logger.log('[Replay] Using simple buffer');
+  return new EventBufferArray();
+}
